Erlang 24 is out, Erlang 27 is (conditionally) in
We can start with just RPMs published to GitHub
releases.

Updating package repositories should not affect
RabbitMQ installations but might affect other
installations, and we know for a fact that our
packages are used without RabbitMQ because
no other options are available on, say, Debian.

Starting with RabbitMQ 4.0.4 Erlang 27 will be
marked as supported so this restriction can
be lifted eventually.