Regenerate workflows to follow Team RabbitMQ's infra changes
namely that we won't use Cloudsmith going forward.

Cloudsmith is an excellent services that we are migrating
away from for reasons entirely unrelated to its quality.

Note that now that we no longer publish Erlang 28
to Cloudsmith, we can at least build the packages
and produces GitHub releases in the rabbitmq/erlang-rpm
repository.