Stop updating Erlang 26.2.x RPM packages
The reasoning largely matches that of a9a409add599
except that maintaining our zero dependency
stripped down Erlang RPM is *significantly*
less time consuming even though we build or support
building on more RPM-based distributions.

But if we stop building 26.x Debian packages,
it'd be odd to continue with the RPM ones.